# Add a benchmark to test the performance of generating the most common values on different locales